Add an EndpointRateLimiter.
Placing the endpoint rate limiter before the others (before the IP-address and login username rate limiters), to prioritize protection against brute-forcing. Hopefully the other rate limiters will not be shadowed by it in the normal case, and allow for some logging of the lead-up scenario with enough logged data to triage and deal with brute-force offenders through IP blocking, etc.