Prepare for invoice crediting.
And the data round-tripping, back from an invoice to a draft.

The reason for this is to avoid having to resort to brittle heuristics of finding a suitable Article, which may have changed since the invoice was issued. So instead of having the draft row rely heavily on the Article, making the draft row more independent of the article, and also more like the invoice row.

Note: bookkeeping account properties may need to be added later, to InvoiceDraftRow.